+++ pylana/shiny_dashboards.py
@@ -1,3 +1,8 @@
+"""
+shiny dashboard management api requests
+"""
+
+
 import io
 import json
 import os
