+++ pylana/structures.py
@@ -1,3 +1,7 @@
+"""
+data classes
+"""
+
 from dataclasses import dataclass
 from typing import List
