+++ pylana/v1.py
@@ -1,3 +1,7 @@
+"""
+legacy lana api
+"""
+
 import json
 import re
 import zipfile
