+++ pylana/v2.py
@@ -1,3 +1,7 @@
+"""
+lana combined api
+"""
+
 from pylana.logs import LogsAPI
 from pylana.resources import ResourceAPI
 from pylana.shiny_dashboards import ShinyDashboardAPI
